fix(ui): clear the remaining react-doctor findings
Ignore *.stories.tsx entirely instead of two rule-scoped exceptions.
Move non-component exports out of columns-menu.tsx and query-bar.tsx
(only-export-components), swap array .includes() lookups in render
loops for Set.has(), lazy-init two useState calls, combine chained
filter/map passes into single reduce()s, and fix a missing useMemo
dependency. Split QueryBar's ~500-line body into SuggestionRow/
SuggestionsPopup plus a use-query-bar hook so it clears
no-giant-component; behavior is unchanged, all 142 tests still pass.
