+++ packages/ui/src/index.ts
@@ -179,15 +179,13 @@ export { useDataTable } from './components/data-table/use-data-table';
 
 /* --- Query bar ----------------------------------------------------------- */
 
-export type {
-  QueryBarProps,
-  QueryField,
-} from './components/query-bar/query-bar';
+export type { QueryBarProps } from './components/query-bar/query-bar';
+export { QueryBar } from './components/query-bar/query-bar';
+export type { QueryField } from './components/query-bar/query-bar-parts';
 export {
-  QueryBar,
   queryFieldsFromColumns,
   variantsFromFields,
-} from './components/query-bar/query-bar';
+} from './components/query-bar/query-bar-parts';
 
 /* --- Shell --------------------------------------------------------------- */
